refactor: 기존의 노드 플로우 구현 삭제
