+++ frontend/src/components/projects/node-flow/NodeMenu.tsx
@@ -1,4 +1,3 @@
-import { useState } from 'react';
 import {
   IconButton,
   Menu,
@@ -15,6 +14,7 @@ import {
   IconPlus,
   IconTrash,
 } from '@wanteddev/wds-icon';
+import { useState } from 'react';
 import { CustomMenuItem } from '@/components/commons/custom-menu/CustomMemuItem';
 
 interface NodeMenuProps {
